Move typed_data directory up
