+++ starknet_py/net/account/account_client.py
@@ -47,7 +47,7 @@ from starknet_py.utils.data_transformer.execute_transformer import (
 )
 from starknet_py.utils.iterable import ensure_iterable
 from starknet_py.utils.sync import add_sync_methods
-from starknet_py.utils.typed_data.typed_data import TypedData as TypedDataDataclass
+from starknet_py.utils.typed_data import TypedData as TypedDataDataclass
 from starknet_py.net.models.typed_data import TypedData
