+++ starknet_py/net/signer/stark_curve_signer.py
@@ -19,7 +19,7 @@ from starknet_py.net.models import (
 )
 from starknet_py.net.signer.base_signer import BaseSigner
 from starknet_py.utils.crypto.facade import message_signature
-from starknet_py.utils.typed_data.typed_data import TypedData as TypedDataDataclass
+from starknet_py.utils.typed_data import TypedData as TypedDataDataclass
 from starknet_py.net.models.typed_data import TypedData
